updates conflicts

+++ app/controllers/api/capstones_controller.rb
@@ -27,17 +27,37 @@ class Api::CapstonesController < ApplicationController
 
   def update
     @capstone = Capstone.find_by(id: params[:id])
-    @capstone.name = params[:name] || @capstone.name
-    @capstone.description = params[:description] || @capstone.description
-    @capstone.url = params[:url] || @capstone.url
-    @capstone.screenshot_url = params[:screenshot_url] || @capstone.screenshot_url
+    if @capstone.student_id == current_student.id
+      @capstone.name = params[:name] || @capstone.name
+      @capstone.description = params[:description] || @capstone.description
+      @capstone.url = params[:url] || @capstone.url
+      @capstone.screenshot_url = params[:screenshot_url] || @capstone.screenshot_url
+      if @capstone.save
+        render "show.json.jb"
+      else
+        render json: { errors: @capstone.errors.full_messages }, status: :unprocessable_entity
+      end
+
+    end
+    
     render "show.json.jb"
+
+    
+
+
+
+
+
   end
   
   def destroy
     capstone = Capstone.find_by(id: params[:id])
-    capstone.destroy
-    render json: {message: "Capstone successfully destroyed."}
+    if capstone.student_id == current_student.id
+      capstone.destroy
+      render json: { message: "Successfully Destroyed Capstone" }
+    else
+      render json: { error: "Not your Capstone to delete!" }, status: :unauthorized
+    end
   end
 
 end
