+++ app/controllers/api/skills_controller.rb
@@ -25,15 +25,27 @@ class Api::SkillsController < ApplicationController
   
   def update
     @skill = skill.find_by(id: params[:id])
-    @skill.current_student_id = params[:current_student_id] || @skill.current_student_id
-    @skill.skill_name = params[:skill_name] || @skill.skill_name
-    render "show.json.jb"
+    if @skill.student_id == current_student.id
+      @skill.skill_name = params[:skill_name] || @skill.skill_name
+      if @skill.save
+        render "show.json.jb"
+      else
+        render json: { errors: @skill.errors.full_messages }, status: :unprocessable_entity
+      end
+    else
+      render json: { error: "Not your skill to update!" }, status: :unauthorized
+    end
+
   end
   
   def destroy
     skill = Skill.find_by(id: params[:id])
-    skill.destroy
-    render json: {message: "Skill successfully destroyed."}
+    if skill.student_id == current_student.id
+      skill.destroy
+      render json: { message: "Successfully Destroyed Skill" }
+    else
+      render json: { error: "Not your Skill to delete!" }, status: :unauthorized
+    end
   end
 
 end
